Add submodule `vim-apathy'
Prior to this change I used vim's path setting as a kind of file-finder.
But doing this breaks a number of other features, as Tim Pope explains
herez;

https://github.com/tpope/vim-apathy

This change removes my path option setting and installs the vim-apathy
plugin, which takes care of setting the path option to a useful value
according to the file type.

As for making the finding of files easier, Tim says the right way is to
use a file-finding plugin such as fzf.