Add more keyboard mappings
Swaps Caps Lock and Escape, and makes both Menu and Win_R behave as
Compose if pressed alone and as Alt (Meta) if pressed with another key.
I've applied this latter mapping to both Win_R and Menu because most
keyboards I use have only one or the other of them.